Moved M.Diagonalizer into Eigen, added Eigen tests
